Add files via upload
V1 of the bot in c#, much work needs to be done especially on the prompt and on giving the bot the context of messages. I also want to add in a filtering system which will prevent the bot attempting to reply to spam, and create certain requirements for a prompt to be sent to the LLM (e.g. messages with less than a few words will be ignored, ect.) at some point I may also add a hash table which holds the previous responses to 1000  of the most common previous messages that have already been sent, allowing for fast lookup and cheap replying to messages that dont require a whole fresh AI prompt)